total_paid is not float is decimal in database

+++ db/migrate/20121019185517_add_total_paid_to_backers.rb
@@ -1,5 +1,5 @@
 class AddTotalPaidToBackers < ActiveRecord::Migration
   def change
-    add_column :backers, :total_paid, :float, :defaul => 0
+    add_column :backers, :total_paid, :decimal, :defaul => 0
   end
 end
